@@ -8,7 +8,7 @@
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
         <version>3.4.7</version>
-        <relativePath/> <!-- lookup parent from repository -->
+        <relativePath/> <!-- Do NOT look for a local parent file Fetch it directly from Maven repositories -->
     </parent>
 
     <groupId>com.example</groupId>
@@ -18,8 +18,7 @@
     <description>Demo project for Spring Boot</description>
 
     <properties>
-        <java.version>17</java.version>
-        <lombok.version>1.18.34</lombok.version> <!-- Updated Lombok version -->
+        <java.version>21</java.version>
     </properties>
 
     <dependencies>
@@ -47,6 +46,12 @@
             <artifactId>spring-boot-starter-security</artifactId>
         </dependency>
 
+        <!-- Spring Boot Starter Validation includes everything you need -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-validation</artifactId>
+        </dependency>
+
         <!-- PostgreSQL driver -->
         <dependency>
             <groupId>org.postgresql</groupId>
@@ -58,8 +63,8 @@
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-            <version>${lombok.version}</version>
-            <scope>provided</scope>
+            <version>1.18.42</version>
+            <scope>compile</scope>
         </dependency>
 
         <!-- Testing -->
@@ -133,8 +138,7 @@
 
     <build>
         <plugins>
-            <!-- Compiler plugin with Lombok annotation processing -->
-            <plugin>
+           <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.11.0</version>
@@ -142,21 +146,18 @@
                     <source>${java.version}</source>
                     <target>${java.version}</target>
                     <annotationProcessorPaths>
-                        <!-- Lombok must come FIRST -->
                         <path>
                             <groupId>org.projectlombok</groupId>
                             <artifactId>lombok</artifactId>
-                            <version>${lombok.version}</version>
+                            <version>1.18.42</version>
                         </path>
-                        
-                        <!-- MapStruct Processor -->
                         <path>
                             <groupId>org.mapstruct</groupId>
                             <artifactId>mapstruct-processor</artifactId>
                             <version>1.6.3</version>
                         </path>
                     </annotationProcessorPaths>
-                    
+        
                     <compilerArgs>
                         <arg>-Amapstruct.defaultComponentModel=spring</arg>
                     </compilerArgs>
