@@ -9,8 +9,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.CacheEvict;
@@ -18,66 +16,50 @@ import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.cache.annotation.Caching;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.example.demo.Main;
 import com.example.demo.events.UserSignUp;
 import com.example.demo.mappers.UserMapper;
 import com.example.demo.model.AuthToken;
-import com.example.demo.model.MyUserDetails;
 import com.example.demo.model.Notification;
 import com.example.demo.model.Task;
+import com.example.demo.model.Task.Status;
 import com.example.demo.model.User;
 import com.example.demo.model.UserDTO;
 import com.example.demo.repositry.NotificationRepositry;
 import com.example.demo.repositry.TaskRepository;
 import com.example.demo.repositry.UserRepositry;
+import com.example.demo.service.JWTService;
 
 import jakarta.servlet.http.HttpServletRequest;
-
-import com.example.demo.model.Task.Status;
-
-
+import jakarta.transaction.Transactional;
 
 
 @Service
 public class UserService {
 
-    
     @Autowired
     UserRepositry userRepositry;
-
-    @Autowired
-    JWTService jwtService;
-
-    @Autowired
-    AuthenticationManager authenticationManager;
     @Autowired
     TaskRepository taskRepository;
     @Autowired
     NotificationRepositry notificationRepositry;
-    // @Autowired(required= false)
-    // JavaMailSender javaMailSender;
-
     @Autowired
     UserMapper mapper;
-
     @Autowired
     ApplicationEventPublisher eventPublisher;
-
     @Autowired
-    MyUserDetailsService myUserDetailsService;
+    AuthService authService;
+    @Autowired
+    JWTService jwtService;
 
-    private static final Logger logger=LogManager.getLogger(Main.class);
+
+    // private static final Logger logger=LogManager.getLogger(Main.class);
 
     BCryptPasswordEncoder encoder= new BCryptPasswordEncoder(10);
+
 
     @Cacheable(value = "users")
     public List<UserDTO> getUsers() {
@@ -100,19 +82,19 @@ public class UserService {
         if (verifyUserNameExist(user.getName())) {
             userRepositry.save(user);
             eventPublisher.publishEvent(new UserSignUp(mapper.toDto(user)));
-            return login(user, request);
+            return authService.login(user, request);
         } else {
             throw new RuntimeException("Username already exists");
         }
     }
 
     @Caching(
-        put = @CachePut(value = "usersById", key = "#result.id"),  // Updates users cache
-        evict = @CacheEvict(value = "users", allEntries = true)  // Clears list cache
+        put = @CachePut(value = "usersById", key = "#result.id"),
+        evict = @CacheEvict(value = "users", allEntries = true)
     )
     public UserDTO updateUser(UserDTO userDTO) {
         User user = userRepositry.findById(userDTO.getId())
-       .orElseThrow(() -> new RuntimeException("User not found"));
+        .orElseThrow(() -> new RuntimeException("User not found"));
     
         user.setName(userDTO.getName());
         user.setEmail(userDTO.getEmail());
@@ -127,11 +109,27 @@ public class UserService {
         return "user deleted!";
     }
 
+    @Transactional
+    public String lockUser(int id){
+        User user = userRepositry.findById(id)
+        .orElseThrow(() -> new RuntimeException("user not found"));
+        user.setLocked(true);
+        userRepositry.save(user);
+        jwtService.revokeUserRefreshTokens(user.getName());
+        return "user locked";
+    }
+
+    @Transactional
+    public String unlockUser(int id){
+        User user = userRepositry.findById(id).orElseThrow(() -> new RuntimeException("user not found"));
+        user.setLocked(false);
+        userRepositry.save(user);
+        jwtService.unrevokeUserRefreshTokens(user.getName());
+        return "user unlocked";
+    }
+
     public boolean verifyUserNameExist(String name) {
-        if (userRepositry.findByName(name) != null) {
-            return false;
-        }
-        return true;
+        return userRepositry.findByName(name) == null;
     }
 
 
@@ -139,13 +137,13 @@ public class UserService {
     public List<Task> getTasks(String name) {
         User user = userRepositry.findByName(name);
         if (user != null) {
-            return user.getTasks();  
+            return user.getTasks();
         }
         return List.of();
     }
 
 
-     public Task getTask(String name, int id) {
+    public Task getTask(String name, int id) {
         User user = userRepositry.findByName(name);
         if (user != null) {
             return user.getTasks().stream()
@@ -156,12 +154,12 @@ public class UserService {
         return null;
     }
 
-    public List<Task> getTasksByStatus(String name, String status){
+    public List<Task> getTasksByStatus(String name,  Status status){
         User user = userRepositry.findByName(name);
-        if (user != null) {           
+        if (user != null) {
             return user.getTasks().stream()
                 .filter(t -> t.getStatus().equals(status))
-               .collect(Collectors.toList());
+                .collect(Collectors.toList());
         } else {
             return null;
         }
@@ -188,7 +186,7 @@ public class UserService {
             if (notification.getSeen()==false) {
                 unseenNotifications.add(notification);
             }
-        };
+        }
         return unseenNotifications;
     }
     
@@ -210,7 +208,7 @@ public class UserService {
         return number;
     }
 
-    private Path imagePath;
+    private final Path imagePath;
 
     public UserService(@Value("${file.upload-dir}") String imagePath) throws IOException{
         this.imagePath=Paths.get(imagePath).toAbsolutePath().normalize();
