@@ -1,10 +1,8 @@
 package com.example.demo.model;
 
 import java.util.Collection;
-import java.util.Collections;
 
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 public class MyUserDetails implements UserDetails {
@@ -39,6 +37,5 @@ public class MyUserDetails implements UserDetails {
     @Override public boolean isCredentialsNonExpired() { return true; }
     @Override public boolean isEnabled() { return true; }
     
-    // Adapts your User entity to Spring Security’s UserDetails interface.
-    // Security doesn’t work directly with your User entity → it needs a standard format (UserDetails).
+
 }
